moved lassen to new peak syntax. Fixed lut test
