+++ tests/test_lut.py
@@ -1,14 +1,10 @@
 from collections import namedtuple
 import operator
 import lassen.asm as asm
-from lassen.sim import gen_pe
-from lassen.isa import DATAWIDTH
+from lassen import PE, Data, DATAWIDTH
 from hwtypes import SIntVector, UIntVector, BitVector, Bit
 import pytest
 import random
-
-Bit = Bit
-Data = BitVector[DATAWIDTH]
 
 op = namedtuple("op", ["name", "func"])
 NTESTS = 32
@@ -19,7 +15,7 @@ NTESTS = 32
     op('xor',  lambda x, y: x^y),
 ])
 def test_lut_binary(op):
-    pe = gen_pe(BitVector.get_family())()
+    pe = PE()
     inst = getattr(asm, f"lut_{op.name}")()
     for _ in range(NTESTS):
         b0 = Bit(random.choice([0,1]))
@@ -33,7 +29,7 @@ def test_lut_binary(op):
     op('not', lambda x: ~x),
 ])
 def test_lut_unary(op):
-    pe = gen_pe(BitVector.get_family())()
+    pe = PE()
     inst = getattr(asm, f"lut_{op.name}")()
     for _ in range(NTESTS):
         b0 = Bit(random.choice([0,1]))
@@ -47,7 +43,7 @@ def test_lut_unary(op):
     op('mux', lambda sel, d0, d1: d1 if sel else d0),
 ])
 def test_lut_ternary(op):
-    pe = gen_pe(BitVector.get_family())()
+    pe = PE()
     inst = getattr(asm, f"lut_{op.name}")()
     for _ in range(NTESTS):
         sel = Bit(random.choice([0,1]))
@@ -58,7 +54,7 @@ def test_lut_ternary(op):
         assert res_p==op.func(sel,d0,d1)
 
 def test_lut():
-    pe = gen_pe(BitVector.get_family())()
+    pe = PE()
     for _ in range(NTESTS):
         lut_val = BitVector.random(8)
         inst = asm.lut(lut_val)
